# swagger/api/swagger.yaml
---
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Logistics"
  contact:
    email: "ad.popov@icloud.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "ligisticstestapp.herokuapp.com"
basePath: "/"
tags:
- name: "requests"
  description: "Everything about requests"
schemes:
- "https"
- "http"
paths:
  /requests:
    get:
      tags:
      - "requests"
      summary: "Get all request"
      operationId: "requestsGET"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Request"
        "404":
          description: "Requests not found"
      x-swagger-router-controller: "Requests"
    post:
      tags:
      - "requests"
      summary: "Add a new request"
      operationId: "requestsPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Request"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Requests"
  /requests/{requestId}:
    get:
      tags:
      - "requests"
      summary: "Find request by ID"
      description: "Returns a single request"
      operationId: "requestsRequestIdGET"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "requestId"
        in: "path"
        description: "ID of request to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Request"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Request not found"
      x-swagger-router-controller: "Requests"
    put:
      tags:
      - "requests"
      summary: "Updates a reguest with form data"
      description: ""
      operationId: "requestsRequestIdPUT"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "requestId"
        in: "path"
        description: "ID of request to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/Request"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Request"
        "400":
          description: "Invalid request"
      x-swagger-router-controller: "Requests"
    delete:
      tags:
      - "requests"
      summary: "Delete request"
      operationId: "requestsRequestIdDELETE"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "requestId"
        in: "path"
        description: "ID of request to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Request"
        "400":
          description: "Invalid request"
      x-swagger-router-controller: "Requests"
definitions:
  Request:
    type: "object"
    properties:
      id:
        type: "string"
      client_full_name:
        type: "string"
      application_date:
        type: "string"
        format: "date"
      carrier_full_name:
        type: "string"
      mobile_number:
        type: "string"
      comments:
        type: "string"
      company:
        type: "string"
      ATI:
        type: "number"
      status:
        type: "string"
        description: "Order Status"
        enum:
        - "placed"
        - "approved"
        - "delivered"
    example:
      comments: "comments"
      client_full_name: "client_full_name"
      ATI: 0.80082819046101150206595775671303272247314453125
      carrier_full_name: "carrier_full_name"
      application_date: "2000-01-23"
      company: "company"
      id: "id"
      mobile_number: "mobile_number"
      status: "placed"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"